Agregar Mantenimiento a Una Orden Hecho Miercoles 4:00

+++ GestionDeTaller.Models/OrdenesDeMantenimiento.cs
@@ -47,6 +47,6 @@ namespace GestionDeTaller.Models
         public int DiasEnProceso { get; set; }
 
         [NotMapped, Display(Name = "Días Trabajados")]
-        public int DiasTrabajados { get; set; }
+        public decimal DiasTrabajados { get; set; }
     }
 }
